+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Developer-side changes
 - Set up bump2version to help with version number management.
 - Raised test coverage to 100%.
 - Upgraded dependency pinning system for cleaner environment setup and update.
+- `pinttr.interpret_units()`: Support for `pint.Quantity` magnitude values.
 
 1.0.0 (2021-02-04)
 ------------------